pyproject.toml
[project]
name = "suprsend-agents-toolkit"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.13.3",
    "pydantic==2.12.5",
    "pyyaml==6.0.2",
    "suprsend-py-sdk @ git+https://github.com/suprsend/suprsend-py-sdk.git@tool-apis",
    "jsonpath-ng==1.8.0",
]

[project.optional-dependencies]
langchain = ["langchain-core==1.2.18"]
mcp = ["mcp==1.26.0"]
jsonnet = ["jsonnet==0.21.0"]

[build-system]
requires = ["setuptools==82.0.1", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
